Move OpenTracing's Tracer API out of Instana::Tracer
