+++ lib/instana/open_tracing/carrier.rb
